added routes and appropriate files
